// src/features/my/model/my.ts
export type MyPost = {
  id: string;
  title: string;
  subtitle: string;
  nickname: string;
  createdAt: string;
  commentCount: number;
};

export type MyProfile = {
  nickname: string;
  introduction: string;
  email: string;
  password: string;
  passwordConfirm: string;
  name: string;
  birthDate: string;
};